mobile css finished
