+++ src/App.jsx
@@ -1,3 +1,9 @@
+import ResultsPage from './pages/ResultPage';
+
 export const App = () => {
-  return <div></div>;
+  return (
+    <div>
+      <ResultsPage />
+    </div>
+  );
 };
