+++ src/components/Results/results.module.css
@@ -1,0 +1,102 @@
+.border {
+    margin-left: auto;
+    margin-right: auto;
+    margin-bottom: 18px;
+    padding-bottom: 16px;
+    width: 130px;
+    border-bottom: solid rgba(0, 0, 0, 0.15) 1px ;
+    
+    
+}
+
+.description {
+    font-family: 'Montserrat';
+        font-style: normal;
+        font-weight: 500;
+        font-size: 10px;
+        line-height: 1.2;
+        text-align: center;
+        letter-spacing: 0.02em;
+        text-transform: uppercase;
+        color: black;
+        opacity: 1 important!;
+}
+
+.list {
+    display: flex;
+    justify-content: center;
+    margin-bottom: 20px;
+    align-items: center;
+}
+
+.graffic {
+    margin-bottom: 20px;
+}
+
+.item {
+    margin: 0 10px;
+    font-family: 'Montserrat';
+        font-style: normal;
+        font-weight: 500;
+        font-size: 10px;
+        line-height: 1.2;
+        letter-spacing: 0.02em;
+}
+
+.decoration {
+    font-family: 'Montserrat';
+        font-style: normal;
+        font-weight: bold;
+        font-size: 20px;
+        line-height: 1.2;
+        letter-spacing: 0.02em;
+        opacity: 0.15;
+        color: #00122F
+}
+
+.number {
+    font-weight: 700;
+}
+
+.image {
+    margin: 0 auto 20px auto;
+}
+
+.mainMessage {
+    margin-bottom: 18px;
+    font-family: 'Montserrat';
+        font-style: normal;
+        font-weight: 700;
+        font-size: 16px;
+        line-height: 1.25;
+        text-align: center;
+        letter-spacing: 0.02em;
+}
+
+.secondaryMessage {
+    margin-bottom: 20px;
+    font-family: 'Montserrat';
+        font-style: normal;
+        font-weight: 500;
+        font-size: 10px;
+        line-height: 1.2px;
+        text-align: center;
+        letter-spacing: 0.02em;
+}
+
+.button {
+    display: block;
+    margin: 0 auto;
+    cursor: pointer;
+        font-family: inherit;
+        background-color: var(--accent-color);
+        padding: 17px 28px;
+        height: 48px;
+        min-width: 190px;
+        color: var(--white-color);
+        font-weight: 400;
+        font-size: 14px;
+        line-height: 1.15;
+        border: 0;
+}
+
