+++ src/pages/ResultPage/index.js
@@ -1,1 +1,1 @@
-export { default } from './ResultPage';
+export { default } from './ResultsPage';
